feat(engine): port dflash2 and upstream a16b644285 onto the sm_89 branch
Three-way merge against upstream main (base a2761ec1e6) with the branch's
sm_89/4080 SUPER contract preserved.

DFlash2 (Qwen3.8-27B): masked-draft proposal, target verification with
configurable K=1..15, candidate selector, 2048 cyclic context ring,
context KV materialization, fused rmsnorm+rope, dynamic grouped conv,
linear top-k, and the converter --dflash2-model companion-weight path.

Branch contracts kept: causal_cache attention family with the rotated/E8
KV codecs and DType dispatch, console_log/load_progress serve redesign,
slot session persistence, resident_resources runtime semantics.

Upstream taken: sliding-window 2048/4096 sharing, dflash2 op suite,
kv-coverage prefix_execution_splits fix, engine is_available, rmsnorm
prefetch/FixedD, sparse_moe small-T S2, gdn gating residency API, and the
httplib 0.54.1 vendored copy (ThreadPool signature adapted; context ring
V storage restored to the FP16 contract across the cyclic cache, prefix
append, and context attention).

sm_89 build guards: NINFER_SM86 branches keep w8 static staging under the
48 KiB per-CTA limit (4-warp tiles at 40+ columns, BK64 fused-swiglu
rowsplit, 8-warp 35B table), and the causal resolver routes batch-one
verify widths up to 12 through chunked small_t (+3.5% DFlash2 K=7 code
decode on RTX 4080 SUPER).

+++ src/ops/softmax_attention/dense/context/context_softmax_attention.cpp
@@ -51,7 +51,7 @@ std::uint32_t validate_context(const PagedKVBatchLayerView& context, const char*
         context.block_tables.ne[0] <= 0 || context.block_tables.ne[1] <= 0) {
         throw std::invalid_argument(std::string(op) + ": invalid context capacity");
     }
-    if (context.k_pages.dtype != DType::BF16 || context.v_pages.dtype != DType::BF16) {
+    if (context.k_pages.dtype != DType::BF16 || context.v_pages.dtype != DType::FP16) {
         throw std::invalid_argument(std::string(op) + ": context K/V must be BF16");
     }
     require_shape(context.k_pages, kHeadDim, kPagedKVPageSize, physical_pages, kKVHeads, op,
